Fix checking localPlayer to be photonView.isMine
Since localPlayer is static, it's always true to check if it's null or
not null for every playercontroller -- for every connected player prefab
in the game. Replaced it with a slightly clearer property to check
photonView.isMine, so now the methods will only continue when they're on
the main player.