Fix Attempt #2 for rubberbanding on SetParent
-Other Changes-
Updated Unity(oops)
Got rid of "win" variable, should be handled internally, and then communicated with an RPC call that the game is over
Got rid of "room properties game active" for the minigame, also can be handled with a simple variable and an RPC call
Added List<GameObject> Blocks for minigameBlockStack, to keep references of the blocks so we can delete them when the minigame is over.
Changed Instantiate to InstantiateSceneObject for minigameBlockStack, so the blocks are there until either the game ends or the last player leaves. (master client deletes when game is done)
Introduced constant for the room name string literal, so people can have a private room and temporarily change it easily when working with their own codebase.